# feat: Add dynamic MCP tools for runtime SQL execution and tool creation